+++ changelog.md
@@ -3,3 +3,4 @@
 + [ENH] : Include `react-router-dom` to manage routes
 + [BUG] : Include app folder
 + [BUG] : Back to original logo image
++ [BUG] : save pylib scripts